refactor(controller): split settings.ts into a barrel over settings/
settings.ts had grown to 4889 lines covering seven unrelated concerns.
It's now the public barrel — it keeps load() and update(), the two
operations that touch the settings file, and re-exports everything else
from ./settings/:

  vocab.ts       fixed value sets, bounds, seed data, pure coercers
  defaults.ts    DEFAULTS + BOUNDS
  store.ts       the loaded-settings cache and its accessors
  normalize.ts   lenient load-path normalizers
  validate.ts    strict update() validators
  persona.ts     persona / show resolution and the prompt fragments
  liquidsoap.ts  the liquidsoap_*.txt writers

Same barrel-over-parts shape llm/ already uses, so no call site changes —
every `from './settings.js'` import keeps working.

The one non-mechanical change is the cache seam: an imported `let` can't
be assigned, so store.ts now exposes peek()/setCache() and settings.ts
writes through those. peek() returns the raw nullable cache, preserving
the pre-load behaviour of the handful of reads that distinguished null
from loaded (renderDjPrompt, resolveOnAirLocation, agentPersonaPreamble)
rather than folding them into get()'s defaults substitution.

No behaviour change. Controller lint clean, all 52 test files pass.
